requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
filelock==3.9.0
Flask==2.2.3
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.72
packaging==23.0
paho-mqtt==1.6.1
pandas==1.5.3
Pillow==9.3.0
psutil==5.9.4
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sympy==1.11.1
torch==2.0.0+cpu
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tqdm==4.65.0
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.3
zipp==3.13.0
